requirements.txt
black~=23.10.0
pygments>=2.7.4
markdown>=3.3
matplotlib
jinja2
requests
mdx_truly_sane_lists
PyYAML
pandas
jupyter
tensorflow~=2.14.0
tensorflow_datasets
keras-core==0.1.7
keras-tuner==1.4.5
keras-cv==0.6.4
keras-nlp==0.6.2
pycocotools
